requirements.txt
streamlit
matplotlib
numpy
opencv-python-headless
tqdm
stqdm
moviepy

Pillow
scikit_learn

networkx
plotly

git+https://github.com/KausikN/StreamlitCommonUtils.git[color,dataset,graph,image,plot,video]